check if BinStore is null (this can happen if a product is corrupt) (cherry-picked from ba11b3a)

+++ beam-binning/src/main/java/org/esa/beam/processor/binning/database/AbstractBinDatabase.java
@@ -48,11 +48,15 @@ abstract public class AbstractBinDatabase implements BinDatabase {
     protected Logger logger;
 
     public void flush() throws IOException {
-        store.flush();
+        if (store != null) {
+            store.flush();
+        }
     }
 
     public void delete() throws IOException {
-        store.delete();
+        if (store != null) {
+            store.delete();
+        }
     }
 
     /**
@@ -100,7 +104,11 @@ abstract public class AbstractBinDatabase implements BinDatabase {
     }
 
     public String getStorageType() {
-        return store.getClass().getName();
+        if (store != null) {
+            return store.getClass().getName();
+        } else {
+            throw new IllegalStateException("Bin store is null. No storage type available");
+        }
     }
 
     /**
@@ -128,8 +136,8 @@ abstract public class AbstractBinDatabase implements BinDatabase {
      */
     protected int sumVarsPerBin() {
         int result = 0;
-        for (int bandIndex = 0; bandIndex < numVarsPerBin.length; bandIndex++) {
-            result += numVarsPerBin[bandIndex];
+        for (int aNumVarsPerBin : numVarsPerBin) {
+            result += aNumVarsPerBin;
         }
         return result;
     }
